+++ src/main/java/top/theillusivec4/caelus/mixin/MixinPlayerEntity.java
@@ -17,7 +17,7 @@
  * License along with Caelus.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package top.theillusivec4.caelus.loader.mixin;
+package top.theillusivec4.caelus.mixin;
 
 import net.minecraft.entity.attribute.DefaultAttributeContainer.Builder;
 import net.minecraft.entity.player.PlayerEntity;
@@ -27,19 +27,20 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import top.theillusivec4.caelus.api.CaelusApi;
-import top.theillusivec4.caelus.loader.common.MixinHooks;
+import top.theillusivec4.caelus.common.util.CommonCaelusHooks;
 
+@SuppressWarnings("unused")
 @Mixin(value = PlayerEntity.class, priority = 100)
-public class PlayerEntityMixin {
+public class MixinPlayerEntity {
 
   @Inject(at = @At("RETURN"), method = "createPlayerAttributes() Lnet/minecraft/entity/attribute/DefaultAttributeContainer$Builder;")
   private static void createAttributes(CallbackInfoReturnable<Builder> cb) {
-    cb.getReturnValue().add(CaelusApi.ELYTRA_FLIGHT);
+    cb.getReturnValue().add(CaelusApi.getInstance().getFlightAttribute());
   }
 
   @Inject(at = @At("TAIL"), method = "tick")
   public void onTick(CallbackInfo cb) {
     @SuppressWarnings("ConstantConditions") PlayerEntity playerEntity = (PlayerEntity) (Object) this;
-    MixinHooks.checkEquippedElytra(playerEntity);
+    CommonCaelusHooks.checkEquippedElytra(playerEntity);
   }
 }